+++ include/tsf/TsfCompartment.h
@@ -13,17 +13,19 @@ using CompartmentChangeCallback = std::function<HRESULT(const GUID *, ULONG)>;
 class TsfCompartment : public ITfCompartmentEventSink
 {
 public:
-    TsfCompartment()                                                    = default;
-    virtual ~TsfCompartment()                                           = default;
+    TsfCompartment() = default;
+
+    virtual ~TsfCompartment() { UnInitialize(); }
+
     TsfCompartment(const TsfCompartment &other)                         = delete;
     TsfCompartment(TsfCompartment &&other) noexcept                     = delete;
     auto operator=(const TsfCompartment &other) -> TsfCompartment &     = delete;
     auto operator=(TsfCompartment &&other) noexcept -> TsfCompartment & = delete;
 
-    auto Initialize(
-        ITfThreadMgr *pThreadMgr, const GUID &guidCompartment, const CompartmentChangeCallback &callback = nullptr
-    ) -> HRESULT;
+    auto Initialize(ITfThreadMgr *pThreadMgr, TfClientId tfClientId, const GUID &guidCompartment, const CompartmentChangeCallback &callback = nullptr)
+        -> HRESULT;
     auto UnInitialize() -> HRESULT;
+    auto SetValue(ULONG value) const -> HRESULT;
     auto GetValue(__out ULONG &pValue) const -> HRESULT;
 
     auto AddRef() -> ULONG override;
@@ -37,6 +39,7 @@ private:
     DWORD                     m_dwCookie = TF_INVALID_COOKIE;
     DWORD                     m_refCount = 0;
     GUID                      m_guidCompartment{};
+    TfClientId                m_tfClientId;
     CompartmentChangeCallback m_callback = nullptr;
 };
 } // namespace Tsf